Tree of queues implemented

+++ src/HydeParkExchange/order.h
@@ -29,7 +29,7 @@ namespace hpx {
 		float price_;
 		int order_id_;
 		int user_id_;
-		order(OrderType type, OrderSide side, int qty, int price) {
+		order(OrderType type, OrderSide side, int qty, float price) {
 			type_ = type;
 			side_ = side;
 			status_ = OrderStatus::Open;
@@ -47,11 +47,11 @@ namespace hpx {
 	};
 
 	struct limit_order : public order {
-		limit_order(OrderSide side, int qty, int price) : order(OrderType::Limit, side, qty, price) {}
+		limit_order(OrderSide side, int qty, float price) : order(OrderType::Limit, side, qty, price) {}
 	};
 
 	struct ioc_order : public order {
-		ioc_order(OrderSide side, int qty, int price) : order(OrderType::IOC, side, qty, price) {}
+		ioc_order(OrderSide side, int qty, float price) : order(OrderType::IOC, side, qty, price) {}
 	};
 }
 #endif // !ORDER_H
